pulto-backend-updated/services/api/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.2
passlib[bcrypt]==1.7.4
pyjwt[crypto]==2.9.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
boto3==1.34.162
minio==7.2.7
paho-mqtt==2.1.0
redis==5.0.7
celery==5.4.0
nbformat==5.10.4
nbclient==0.10.0
orjson==3.10.7
python-magic==0.4.27
